Change Python lexer specification to allow for badly formed decimal numbers (#94)
I added a BADDECNUMBER token to allow malformed literals like 0377 to be detected, which was valid under Python 2 for octal literals but is no longer valid. This will allow better error messages to be provided, though I haven't done this at the parser level with a FAIL because there could be a need for more nuanced messages than FAIL allows. Note that e.g. "00" is a legal decimal literal, but e.g. "01" is not. An example of a more nuanced error message at parser level would be:

* When encountering e.g. 099, use the message "Leading zeros in decimal integer literals are not permitted".
* When encountering e.g. 0377, use the above message but append "; use an 0o prefix for octal integers".