MILESTONE! The CSharp parser parses all of the mono standard class library. That was the initial goal I set out and it is accomplished. Except for 2 (or maybe 3) caveats that will be addressed shortly.
First of all, the preprocessor is implemented somewhat incorrectly. I had
previously thought the preprocessor was purely prelexical and had no
knowledge at all of the structure of a Csharp program. However, this is not
quite so. You can comment out Csharp pre-processor directives with
multiline comments. (Single line comments are irrelevant since a pre-processor
directive has to start a new line anyway.) Also, since there are multiline string
literals, there is the possibility that a line in a string literal could
start with #if or something like that. So a completely correct pre-processor
has to know about both multiline comments and multiline string literals. So,
there were about half a dozen files in the mono distro that relied on the ability
to comment out pp directives, and those didn't parse. But I don't count them
as failures, since the pre-processor is really a separate little mini-grammar.

The same can be said for interpolated strings. I wrote a rather gnarly regexp
but I don't think it is possible to handle interpolated strings purely that way,
since they are allowed to nest, and I don't think a regexp can handle that.
So, finally, interpolated strings are yet another little minigrammar that
has to be implemented separately. There is only one file in the entire mono
distro that requires separate parsing of an interpolated string, so it is
quite a rare case. In practice, people try to avoid writing these overly
complex things, like overly complex regexpes. In any case, both the
pre-processor and the interpolated strings are really separate little mini-grammars
that I will deal with separately, but the main syntactical grammar of C sharp
seems to be implemented.

NB. The other issue is that it is not absolutely clear which version of the language.
I think the parser is probably pretty rock solid for 6.0 and probably most of 7.x.
But I am not sure that the mono source code is a good test suite for 7.x. I'm going to eventually
try to get it to parse all the roslyn source which, I think, uses bleeding edge
language features. But, I need to take a break from this and get to some other
problems now.